separate between tf and np processing
